- Removed safelist property from configs. - Moved mergeConfigs to utilsUI module.
