+++ package.json
@@ -53,7 +53,7 @@
     "@release-it/bumper": "^6.0.1",
     "@vitejs/plugin-vue": "^5.0.5",
     "@vue/eslint-config-prettier": "^9.0.0",
-    "@vueless/plugin-vite": "^0.0.74",
+    "@vueless/plugin-vite": "^0.0.75",
     "@vueless/storybook": "^0.0.37",
     "@vueless/web-types": "^0.0.17",
     "autoprefixer": "^10.4.19",
