+++ src/ui.dropdown-badge/config.js
@@ -30,5 +30,4 @@ export default /*tw*/ {
     /* icons */
     dropdownIcon: "keyboard_arrow_down",
   },
-  safelist: (colors) => [{ pattern: `ring-(${colors})-700` }],
 };
