+++ src/ui.form-radio/config.js
@@ -26,10 +26,4 @@ export default /*tw*/ {
     checked: false,
     disabled: false,
   },
-  safelist: (colors) => [
-    { pattern: `border-(${colors})-500`, variants: ["focus"] },
-    { pattern: `border-(${colors})-800`, variants: ["active"] },
-    { pattern: `text-(${colors})-600`, variants: ["checked"] },
-    { pattern: `ring-(${colors})-700`, variants: ["focus"] },
-  ],
 };
