+++ src/ui.text-badge/config.js
@@ -63,12 +63,4 @@ export default /*tw*/ {
     round: false,
     bordered: false,
   },
-  safelist: (colors) => [
-    { pattern: `bg-(${colors})-50` },
-    { pattern: `bg-(${colors})-600` },
-    { pattern: `border-(${colors})-100` },
-    { pattern: `border-(${colors})-600` },
-    { pattern: `text-(${colors})-600` },
-    { pattern: `ring-(${colors})-700`, variants: ["focus", "focus-within"] },
-  ],
 };
