+++ src/ui.text-header/config.js
@@ -40,5 +40,4 @@ export default /*tw*/ {
     line: true,
     underlined: false,
   },
-  safelist: (colors) => [{ pattern: `text-(${colors})-600` }, { pattern: `border-(${colors})-600` }],
 };
